% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotFCs.R
\name{boxplotFCs}
\alias{boxplotFCs}
\title{Create a box plot for one or more genes across conditions}
\usage{
boxplotFCs(
  fitCofComp,
  gene,
  colors = NULL,
  ylim = NULL,
  las = 2,
  dataPoints = ""
)
}
\arguments{
\item{fitCofComp}{A comparison of fitness coefficients generated by the fitCofComp function.}

\item{gene}{A list of genes to include in the box plot.}

\item{colors}{A list of colors to use for the included genes.}

\item{ylim}{The limits of the y axis.}

\item{las}{The text orientation parameter.}

\item{dataPoints}{A single-letter string specifying the character to use to draw the data points}
}
\description{
This function generates a boxplot of fitness coefficients from an object created with the fitCofComp function.
The set of genes plotted can is specified as a string (for a single gene) or an array of strings (for multiple genes).
Colors can be specified in a similar way (one color per gene in the same order).
}
