% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCofComp.R
\name{fitCofComp}
\alias{fitCofComp}
\title{Compares fitness coefficients between multiple conditions}
\usage{
fitCofComp(
  control,
  treatmentList,
  dataTable = NULL,
  undefCountTable = NULL,
  correction = "undefined"
)
}
\arguments{
\item{control}{The name of the control condition that the test condition should be compared to.}

\item{treatmentList}{A list of treatments (in array format) that should be compared to the control condition.}

\item{correction}{Method for correcting for uneven count distribution between samples (can be "undefined", "median" or "none").}
}
\value{
Returns a fitCofComp object (a matrix with fitness coefficients for every gene and replicate in each tested condition).
}
\description{
This function generates a fitCofComp object, comparing any number of test conditions to the same control condition.
}
