% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGSC.R
\name{generateGSC}
\alias{generateGSC}
\title{Generate a Gene Set Collection}
\usage{
generateGSC(file, columns = c(1, 3), header = FALSE)
}
\arguments{
\item{file}{Tab-separated table containing associations between genes and, e.g., pathways.}

\item{columns}{A pair of columns to be included in the GSC (1 and 3 by default).}

\item{header}{If TRUE, the table file is assumed to have a header line before the data.}
}
\value{
Returns a gene set collection (GSC) to be used with the genesetAnalysis function.
}
\description{
Loads a gene set collection from a tab-separated. The function will use the specified columns, by default column 1 and 3. The first column will be assumed to be gene names and the other column will be assumed to contain gene sets, e.g. pathways. By default, no header row is assumed in the input file.
}
