% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetAnalysis.R
\name{genesetAnalysis}
\alias{genesetAnalysis}
\title{Performs a gene set analysis}
\usage{
genesetAnalysis(GSC, dataset, p = 0.05, test = "t.test", center = FALSE)
}
\arguments{
\item{GSC}{Gene Set Collection object fenerateed by the generateGSC function.}

\item{dataset}{A data set generated by the rankByP function.}

\item{p}{Cutoff p-value for inclusion in the gene set analysis.}

\item{test}{Type of statistical test to use (default is "t.test", but can also be "wilcox").}

\item{center}{If TRUE, the mean fitness coefficient of each sample is subtracted from all values in that sample before analysis.}
}
\value{
Returns a table of gene sets (e.g. pathways) and the number of genes in each set with a positive fitness coefficient, a negative fitness coefficient, the number of non-significant genes, the estimated average fitness coefficient in the set, the uncorrected p-value for the enrichment, and the FDR corrected p-value for enrichment.
}
\description{
Performs a gene set analysis based on a supplied GSC (see generateGSC function) and a data set produced by the rankByP function.
By default, only genes with a significantly different fitness coefficient (p < 0.05) are included in the gene set analysis. This can be changed by changing the significance cutoff to any value between 0 and 1.
On each gene set, a statistical test of differential sum of fitness coefficients is performed.
Be default, this is a t-test, but this can be changed to either “t.test” or “wilcox” (for the Wilcoxon signed rank test).
If center is set to TRUE, the mean fitness coefficient of each sample is subtracted from that sample before gene set analysis is performed. By default, this centration of data is not performed.
}
