% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadData}
\alias{loadData}
\title{Loads a TnSeq/INSeq data set}
\usage{
loadData(inputFileName = "")
}
\arguments{
\item{inputFileName}{cafe-annotate output file to use for analysis.}
}
\description{
Loads a specified data set from the cafe-map and cafe-annotate tools (or equivalent) and prepares it for analysis using the CAFE R package.
The function returns nothing, but creates the following objects that can be used by the user (or other functions): countData, undefCounts, unmappedCounts, countMatrix, geneNames, sampleNames, treatmentClasses, treatments
}
