% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFCs.R
\name{plotFCs}
\alias{plotFCs}
\title{Create a scatter plot of how genes compare between two (or more) conditions}
\usage{
plotFCs(
  fitCofList,
  colorList = NULL,
  conditionA = "Fitness coefficient",
  conditionB = "Fitness coefficient",
  lim = c(-11, 11)
)
}
\arguments{
\item{fitCofList}{A list of fitness coefficient objects generated by the fitCof function.}

\item{colorList}{A list of colors to use for the testing conditions.}

\item{conditionA}{Label for the control condition axis.}

\item{conditionB}{Label for the test conditions axis.}

\item{lim}{The limits of the x and y axes.}
}
\description{
This function generates a scatter plot of fitness coefficients from one or more objects created with the fitCof function.
The first condition in the list of fitness coefficient objects will be treated as the control condition.
}
