% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankByP.R
\name{rankByP}
\alias{rankByP}
\title{Perform statistical testing and rank genes}
\usage{
rankByP(
  fitCofTable,
  test = "t.test",
  correction = "BH",
  pcut = 1,
  toplist = NULL
)
}
\arguments{
\item{test}{The type of statistical test to use.}

\item{correction}{The type of correction for multiple testing to use.}

\item{pcut}{Cutoff p-value to be included in output (default = 1, will include all genes).}

\item{toplist}{If specified, the list is restricted to this number of genes. If specified as FALSE, the returned object will not be sorted, but will retain its original order.}

\item{fitCofList}{A fitness coefficient object generated by the fitCof function.}
}
\value{
Returns a list of statistical tests for genes, ranked by p-value.
}
\description{
This function performs statistical testing of all genes in a fitness coefficient object generated by the fitCof function.
By default, the function performs a t.test, followed by Benjamini-Hochberg correction for multiple testing (FDR). Other available tests are "wilcox.test" and "nonlog.t.test"). And method accepted by the R p.adjust function can be used for multiple testing correction.
}
